requirements.txt
streamlit
tensorflow
keras
groq
python-dotenv
faiss-cpu
sentence-transformers
numpy
scikit-learn
